55_HousePrice_Agepredict/Age_prediction/requirements.txt
tensorflow
matplotlib.pyplot
sklearn
numpy
wandb
scipy
Pillow
numpy
pandas
sklearn
opencv-python
dlib